app/requirements.txt
python-dotenv==0.19.0
scrapy==2.5.0
todoist-python==8.1.3
